FBP-193. Basic tests for timer events
