+++ camunda-xml-parser/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/camunda/parser/nodes/process/IntermediateCatchEvent.kt
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.events.catching.BpmnIntermediateCatchingEvent
 import com.valb3r.bpmn.intellij.plugin.camunda.parser.nodes.BpmnMappable
+import com.valb3r.bpmn.intellij.plugin.camunda.parser.nodes.process.nested.definitions.TimerEventDefinition
 import org.mapstruct.Mapper
 import org.mapstruct.factory.Mappers
 
@@ -36,10 +37,6 @@ data class IntermediateCatchEvent(
     interface LinkEventDefinitionMapping {
         fun convertToDto(input: LinkEventDefinition) : BpmnIntermediateCatchingEvent.LinkEventDefinition
     }
-
-    data class TimerEventDefinition(
-            val timeDate: String? = null
-    )
 
     data class SignalEventDefinition(
             @JacksonXmlProperty(isAttribute = true) val signalRef: String? = null
