+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/events/catching/BpmnIntermediateTimerCatchingEvent.kt
@@ -2,17 +2,16 @@ package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.events.catching
 
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.BpmnElementId
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.WithBpmnId
+import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.events.begin.props.BpmnTimerEventDefinition
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.types.IntermediateCatchingEventAlike
 
 data class BpmnIntermediateTimerCatchingEvent(
-        override val id: BpmnElementId,
-        val name: String? = null,
-        val documentation: String? = null,
-        val incoming: List<String>? = null,
-        val outgoing: List<String>? = null,
-        val timeDate: String? = null,
-        val timeDuration: String? = null,
-        val timeCycle: String? = null,
+    override val id: BpmnElementId,
+    val name: String? = null,
+    val documentation: String? = null,
+    val incoming: List<String>? = null,
+    val outgoing: List<String>? = null,
+    val timerEventDefinition: BpmnTimerEventDefinition? = null,
 ) : WithBpmnId, IntermediateCatchingEventAlike {
 
     override fun updateBpmnElemId(newId: BpmnElementId): WithBpmnId {
